refactor: create HeaderRequestHandler
